Move pre-flight checks earlier to start of genGroup
